WIP: Replacing io interface with half-duplex manager methods for sessions
The SessionIO business was just too cumbersome. It is probably going to
be much easier to just return the functions directly, avoiding the need
for extra packages and gets rid of any worries for cyclic dependencies.
Currently a work in progress, new LeaseReceiver and LeaseSender methods
have been added to the manager's main API. These will return a
receiver/sender function as well as a close function.